Relocate swaps logic & types
