+++ packages/transaction-controller/src/constants.ts
@@ -1,5 +1,3 @@
-import { TransactionType } from './types';
-
 export const CHAIN_IDS = {
   MAINNET: '0x1',
   GOERLI: '0x5',
@@ -23,14 +21,6 @@ export const CHAIN_IDS = {
   ARBITRUM: '0xa4b1',
   ZKSYNC_ERA: '0x144',
 } as const;
-
-const SWAPS_TESTNET_CHAIN_ID = '0x539';
-
-
-// An address that the metaswap-api recognizes as the default token for the current network,
-// in place of the token address that ERC-20 tokens have
-export const DEFAULT_TOKEN_ADDRESS =
-  '0x0000000000000000000000000000000000000000';
 
 export const DEFAULT_ETHERSCAN_DOMAIN = 'etherscan.io';
 export const DEFAULT_ETHERSCAN_SUBDOMAIN_PREFIX = 'api';
@@ -113,93 +103,3 @@ export const ETHERSCAN_SUPPORTED_NETWORKS = {
     subdomain: `${DEFAULT_ETHERSCAN_SUBDOMAIN_PREFIX}-gnosis`,
   },
 };
-
-export const SWAP_TRANSACTION_TYPES = [
-  TransactionType.swap,
-  TransactionType.swapApproval,
-];
-
-// Only certain types of transactions should be allowed to be specified when
-// adding a new unapproved transaction.
-export const VALID_UNAPPROVED_TRANSACTION_TYPES = [
-  ...SWAP_TRANSACTION_TYPES,
-  TransactionType.simpleSend,
-  TransactionType.tokenMethodTransfer,
-  TransactionType.tokenMethodTransferFrom,
-  TransactionType.contractInteraction,
-];
-
-export interface SwapsTokenObject {
-  /**
-   * The name for the network
-   */
-  name: string;
-  /**
-   * An address that the metaswap-api recognizes as the default token
-   */
-  address: string;
-  /**
-   * Number of digits after decimal point
-   */
-  decimals: number;
-}
-
-export const ETH_SWAPS_TOKEN_OBJECT: SwapsTokenObject = {
-  name: 'Ether',
-  address: DEFAULT_TOKEN_ADDRESS,
-  decimals: 18,
-};
-
-export const BNB_SWAPS_TOKEN_OBJECT: SwapsTokenObject = {
-  name: 'Binance Coin',
-  address: DEFAULT_TOKEN_ADDRESS,
-  decimals: 18,
-} as const;
-
-export const MATIC_SWAPS_TOKEN_OBJECT: SwapsTokenObject = {
-  name: 'Matic',
-  address: DEFAULT_TOKEN_ADDRESS,
-  decimals: 18,
-} as const;
-
-export const AVAX_SWAPS_TOKEN_OBJECT: SwapsTokenObject = {
-  name: 'Avalanche',
-  address: DEFAULT_TOKEN_ADDRESS,
-  decimals: 18,
-} as const;
-
-export const TEST_ETH_SWAPS_TOKEN_OBJECT: SwapsTokenObject = {
-  name: 'Test Ether',
-  address: DEFAULT_TOKEN_ADDRESS,
-  decimals: 18,
-} as const;
-
-export const GOERLI_SWAPS_TOKEN_OBJECT: SwapsTokenObject = {
-  name: 'Ether',
-  address: DEFAULT_TOKEN_ADDRESS,
-  decimals: 18,
-} as const;
-
-export const ARBITRUM_SWAPS_TOKEN_OBJECT: SwapsTokenObject = {
-  ...ETH_SWAPS_TOKEN_OBJECT,
-} as const;
-
-export const OPTIMISM_SWAPS_TOKEN_OBJECT: SwapsTokenObject = {
-  ...ETH_SWAPS_TOKEN_OBJECT,
-} as const;
-
-export const ZKSYNC_ERA_SWAPS_TOKEN_OBJECT: SwapsTokenObject = {
-  ...ETH_SWAPS_TOKEN_OBJECT,
-} as const;
-
-export const SWAPS_CHAINID_DEFAULT_TOKEN_MAP = {
-  [CHAIN_IDS.MAINNET]: ETH_SWAPS_TOKEN_OBJECT,
-  [SWAPS_TESTNET_CHAIN_ID]: TEST_ETH_SWAPS_TOKEN_OBJECT,
-  [CHAIN_IDS.BSC]: BNB_SWAPS_TOKEN_OBJECT,
-  [CHAIN_IDS.POLYGON]: MATIC_SWAPS_TOKEN_OBJECT,
-  [CHAIN_IDS.GOERLI]: GOERLI_SWAPS_TOKEN_OBJECT,
-  [CHAIN_IDS.AVALANCHE]: AVAX_SWAPS_TOKEN_OBJECT,
-  [CHAIN_IDS.OPTIMISM]: OPTIMISM_SWAPS_TOKEN_OBJECT,
-  [CHAIN_IDS.ARBITRUM]: ARBITRUM_SWAPS_TOKEN_OBJECT,
-  [CHAIN_IDS.ZKSYNC_ERA]: ZKSYNC_ERA_SWAPS_TOKEN_OBJECT,
-} as const;
